S2012

+++ src/main/java/fr/eni/ludotech/dal/ExemplaireRepository.java
@@ -3,5 +3,5 @@ package fr.eni.ludotech.dal;
 import fr.eni.ludotech.bo.Exemplaire;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExemplaireRepository extends JpaRepository<Exemplaire, Long> {
+public interface ExemplaireRepository extends JpaRepository<Exemplaire, Integer> {
 }
