+++ src/main/java/fr/eni/ludotech/dal/JeuRepository.java
@@ -3,5 +3,5 @@ package fr.eni.ludotech.dal;
 import fr.eni.ludotech.bo.Jeu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JeuRepository extends JpaRepository<Jeu, Long> {
+public interface JeuRepository extends JpaRepository<Jeu, Integer> {
 }
